+++ app/__init__.py
@@ -3,5 +3,5 @@ NXZ NLU Service - Natural Language Understanding Service
 逆行者语义识别服务
 """
 
-__version__ = "0.0.2"
+__version__ = "0.0.5"